Feat/initial distrib constructor (#64)
Issue: #60 

This is a first draft. 

In the proposed implementation, the dev would pass an array
`initial_holders` of tuples `(initial_holder_address, amount)`.
The `initial_recipient` should be in the position 0 of the array. As
`initial_recipient` is an `initial_holder`, it makes sense to add him in
the array. But it implies that the sum of all amounts of the
`initial_holders` array should remain equal to `initial_supply`. This is
verified by line 116 `assert(initial_minted_supply <= initial_supply,
'Unruggable: max supply reached');`.
I'm also considering that the `initial_recipient` account should deposit
ALL of his tokens when `launch_memecoin` is called.. Otherwise it is not
possible to perform the "team allocation" check since the balance of
`initial_recipient` would be greater. If `initial_recipient` does not
deposit all of his token in `launch_memecoin` then I'm not performing
the "max team allocation" check on him which bypasses the rule (does
that make sense ?).

Didn't write the tests yet, I wanted to have feedbacks on my
implementation.