redesign initial distribution logic (#78)
Resolves #77.

Also brings some modifications to the `_enforce_holders_limit` internal.
The method can also decrease holders count if the sender has no tokens
left after the transfer.

Lastly, it might be interesting to discuss about adding additional
checks in the `_enforce_holders_limit` such as asserting the `amount`
param or `recipient` are not null. With current ERC20 implementation,
those checks are useless, but it makes this method rely on ERC20
behaviour. Without those checks, this method is a kind of ERC20
extension, which is not a bad design, but also not necessarily what we
want.