Delay that bypasses transfer limit (#83)
The test is not implemented. We need to have a function that
successfully launches memecoin then we will be able to test that.
The test will be simple. Just increase the timestamp by 1000 after
calling `launch_memecoin`

On the factory side, I passed `1000_u64` as the third param of the
constructor of memecoin. It has to be changed by how you gonna use it
from the factory.