audit: disable quote token memecoin (#171)
As reported in the audit by
@ermvrs - [H-01]

The problem was that if
- Someone creates a memecoin token A - Factory has 100% of the supply
- I launch a memecoin B and on Ekubo and I say that I use A as a quote
- Since ekubo launches are not consuming any quote tokens, the rest of A
is transferred to me, the caller. As a result, I now hold the factory's
balance of A - in our example, 100% of A.

After which I can
- transfer let's say 10% of A to factory
- The owner of A calls launch, which will use the balanceOf factory so
10% of the supply
- and it will launch with 10% of the supply in the pool and I will still
hold with 90%

Mitigation:

- Add a check to ensure the no memecoin can be used as quote token.

- never give back the quote tokens that were mistakenly sent to this
contract. Given that there should never be any transfer of quote tokens
to this contract in the case of an Ekubo launch, and that launching on
Jediswap consumes the entirety of the balance of quoteTokens atomically
in a transaction (caller -> factory -> pair), there should never be any
leftovers in the normal flow.